[SC-4521] Give idea capture its own engine, off the ideation session
The Ideas column's `+` reaches CreateIdea through the ideation engine, which
is about to be deleted. Capture uses no session state, so it moves to its own
type with its own Server field and its own route file; the engine method stays
as a delegation for this commit only, so nothing changes behaviour here.
